requirements.txt
autopep8==2.3.2
backports.tarfile==1.2.0
boto3==1.40.18
celery==5.5.3
channels-redis==4.3.0
crispy-bootstrap5==2025.6
daphne==4.2.1
dj-database-url==3.0.1
django-storages==1.14.6
factory-boy==3.3.3
git-filter-repo==2.47.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
pip-chill==1.0.3
platformdirs==4.2.2
psycopg2-binary==2.9.10
pyopenssl==25.1.0
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-sugar==1.1.1
python-dotenv==1.1.1
service-identity==24.2.0
tomli==2.0.1
whitenoise==6.9.0
zope.interface==7.2
